@@ -1,23 +1,20 @@
-[tool.poetry]
+[project]
 name = "kafka-demo"
 version = "0.1.0"
 description = ""
-authors = ["Ashal aka JOKER <ahastudio@gmail.com>"]
 readme = "README.md"
+requires-python = ">=3.13"
+dependencies = [
+    "fastavro>=1.10.0",
+    "kafka-python-ng>=2.2.3",
+]
 
-[tool.poetry.dependencies]
-python = "3.11.5"
-kafka-python = "^2.0.2"
-fastavro = "^1.8.4"
-
-[tool.poetry.group.dev.dependencies]
-isort = "^5.12.0"
-black = "^23.10.0"
-ruff = "^0.1.1"
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
+[dependency-groups]
+dev = [
+    "black>=24.10.0",
+    "isort>=5.13.2",
+    "ruff>=0.8.4",
+]
 
 [tool.black]
 line-length = 79
@@ -27,4 +24,4 @@ profile = "black"
 line_length = 79
 
 [tool.ruff]
-select = ["A", "B", "C"]
+lint.select = ["A", "B", "C"]
